Implemented a possible fix for the __del__ issue
Implemented a possible fix for the __del__ issue, this issue happens on a device running unix based operating systems, i.e. linux or mac, and is an issue with python itself, to avoid it, I have tried using pop instead of del to remove the key:value pair entry of requests. On an attempt to delete the key value pair it used to try deleting the task object as well which was from asyncio, and thats why the exception used to occur, this should hopefully fix it, if it doesnt i'll try other methods.